Tratamento de erros e correção bug encontrado
Foram adicionados tratamento de erro para os métodos de "Update" e "Delete".

Foi identificado e corrigido um erro na classe "ProductDTO", onde foi excluído o construtor padrão da classe, a falta deste causou excessões na tentativa de atualizar e inserir produtos no banco de dados.

O erro foi sanado e testes foram realizados para confirmação deste, assim como confirmação do funcionamento dos tratamentos de erros para os métodos "Update" e "Delete".

Os erro tratado é para quando o usuário tentar atualizar ou deletar um produto cujo o id não exista no banco de dados.

Um tratamento para a tentativa do usuário de deletar um produto que esteja associado a um pedido de compras no banco de dados.

+++ src/main/java/br/com/julianomarthins/dscommerce/services/ProductService.java
@@ -1,41 +1,49 @@
 package br.com.julianomarthins.dscommerce.services;
 
+
 import br.com.julianomarthins.dscommerce.dto.ProductDTO;
 import br.com.julianomarthins.dscommerce.entities.Product;
 import br.com.julianomarthins.dscommerce.respositories.ProductRepository;
+import br.com.julianomarthins.dscommerce.services.exceptions.DatabaseException;
 import br.com.julianomarthins.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+
 
 @Service
 public class ProductService {
 
+    // Atributo
     @Autowired
     private ProductRepository repository;
 
 
     // Retorna ao usuário um produto filtrado pelo id
     @Transactional(readOnly = true)
-    public ProductDTO findById(Long id){
+    public ProductDTO findById(Long id) {
         Product product = repository.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException("Produto não encontrado no banco de dados")
+                () -> new ResourceNotFoundException("Recurso não encontrado")
         );
         return new ProductDTO(product);
     }
 
     // Retorna ao usuário todos os produtos organizados por páginas
     @Transactional(readOnly = true)
-    public Page<ProductDTO> findAll(Pageable pageable){
+    public Page<ProductDTO> findAll(Pageable pageable) {
         Page<Product> result = repository.findAll(pageable);
         return result.map(ProductDTO::new);
     }
 
     // Adiciona um novo produto ao banco de dados
     @Transactional
-    public ProductDTO insert(ProductDTO dto){
+    public ProductDTO insert(ProductDTO dto) {
         Product entity = new Product();
         copyDtoToEntity(dto, entity);
         entity = repository.save(entity);
@@ -44,29 +52,40 @@ public class ProductService {
 
     // Atualiza um produto no banco de dados
     @Transactional
-    public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-       copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+    public ProductDTO update(Long id, ProductDTO dto) {
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
     // Apaga um produto do banco de dados filtrado pelo id
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+
     }
 
 
-
     // Função auxiliar, instancia um novo objeto do tipo Product
-    private void copyDtoToEntity(ProductDTO dto, Product entity){
+    private void copyDtoToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
         entity.setDescription(dto.getDescription());
         entity.setPrice(dto.getPrice());
         entity.setImgUrl(dto.getImg());
     }
-
 
 
 }
